feat(frontend): surface bot liquidations as Explorer Protocol lens
The /explorer/liquidations page exists but wasn't discoverable — no
nav, no overview card. Three surfaces added:

1. **Protocol lens tab "Liquidations"** between Stability Pool and
   Redemptions. Clicking it loads the same content as the standalone
   /explorer/liquidations page (extracted into `LiquidationsLens.svelte`,
   referenced by both surfaces).
2. **Overview card "Bot Liquidations"** below the pool health strip.
   Compact stats (debt covered, collateral received, records, stuck
   claims) with a click-through to the lens. Stuck-claims count turns
   red when non-zero.
3. **Top nav highlighting**: /explorer/liquidations now matches the
   Protocol tab (was Activity) since it's a subsystem dashboard, not
   an event stream.
