+++ src/vault_frontend/src/lib/components/explorer/lenses/lensTypes.ts
@@ -2,6 +2,7 @@ export type LensId =
   | 'overview'
   | 'collateral'
   | 'stability'
+  | 'liquidations'
   | 'redemptions'
   | 'revenue'
   | 'dexs'
